服務佮載入攏照服務分開
