+++ 臺灣言語服務/Moses服務.py
@@ -7,14 +7,12 @@ from 臺灣言語工具.解析整理.拆文分析器 import 拆文分析器
 from 臺灣言語工具.解析整理.文章粗胚 import 文章粗胚
 from 臺灣言語工具.斷詞.拄好長度辭典揣詞 import 拄好長度辭典揣詞
 from 臺灣言語工具.斷詞.語言模型揀集內組 import 語言模型揀集內組
-from 臺灣言語工具.語音合成.語音標仔轉換 import 語音標仔轉換
 
 
-class 服務:
+class Moses服務:
 
-    def __init__(self, 全部翻譯母語模型={}, 全部合成母語模型={}):
+    def __init__(self, 全部翻譯母語模型={}):
         self.全部翻譯母語模型 = 全部翻譯母語模型
-        self.全部合成母語模型 = 全部合成母語模型
 
     def 正規化翻譯支援腔口(self, request):
         全部翻譯母語 = sorted(self.全部翻譯母語模型.keys())
@@ -108,69 +106,6 @@ class 服務:
         except:
             pass
         return self.json包做回應(翻譯結果)
-
-    def 語音合成支援腔口(self, request):
-        return self.json包做回應({'腔口': sorted(self.全部合成母語模型.keys())})
-
-    @csrf_exempt
-    def 語音合成(self, request):
-        if request.method == 'GET':
-            連線參數 = request.GET
-        else:
-            連線參數 = request.POST
-        try:
-            查詢語句 = 連線參數['查詢語句']
-        except:
-            查詢語句 = '你｜li2 好-無｜ho2-0bo5 ？｜? 我｜gua2 足｜tsiok4 好｜ho2 ！｜!'
-        try:
-            查詢腔口 = 連線參數['查詢腔口']
-            合成母語模型 = self.全部合成母語模型[查詢腔口]
-        except:
-            查詢語句 = '你｜li2 好-無｜ho2-0bo5 ？｜? 我｜gua2 足｜tsiok4 好｜ho2 ！｜!'
-            查詢腔口 = '閩南語'
-            合成母語模型 = self.全部合成母語模型[查詢腔口]
-        return self._語音合成實作(合成母語模型, 查詢語句)
-
-    def _語音合成實作(self, 合成母語模型, 查詢語句):
-        母語章物件 = (
-            拆文分析器.分詞章物件(查詢語句)
-            .轉音(合成母語模型['拼音'], 函式='音值')
-        )
-        try:
-            音值物件 = 母語章物件.做(合成母語模型['變調'], '變調')
-        except:
-            音值物件 = 母語章物件
-        標仔陣列 = 語音標仔轉換.物件轉完整合成標仔(音值物件)
-        愛合成標仔 = 語音標仔轉換.跳脫標仔陣列(標仔陣列)
-        if 愛合成標仔 == []:
-            愛合成標仔.append(語音標仔轉換.恬標仔())
-        音檔 = 合成母語模型['模型'].合成(愛合成標仔)
-#         調好音 = self.音標調音(查詢腔口, 音檔)
-        return self.音檔包做回應(音檔.wav格式資料())
-
-#     腔放送進度 = {偏漳優勢音腔口:1.0, 偏泉優勢音腔口:1.0,
-#         混合優勢音腔口:1.0,
-#         四縣腔:1.0, 海陸腔:1.05, 大埔腔:1.6,
-#         饒平腔:1.02, 詔安腔:1.02, }
-#
-#     def 標仔合音檔(self, 查詢腔口, 標仔):
-#         模型 = 全部合成母語模型[查詢腔口]
-#         速度 = self.腔放送進度[查詢腔口]
-#         音檔 = self.轉音檔.合成(模型, 速度, 標仔)
-#         return 音檔
-
-#     def 音標調音(self, 查詢腔口, 音檔):
-#         if 查詢腔口.startswith(閩南語):
-#             調好音 = self.音盒.篩雜訊(音檔)
-#         elif 查詢腔口.startswith(客話):
-#             調好音 = self.音盒.篩懸音(音檔, 6000)
-#         else:
-#             調好音 = 音檔
-#         return 調好音
-
-    def 文字包做回應(self, 文字):
-        回應 = HttpResponse(文字)
-        return 回應
 
     def json包做回應(self, json):
         回應 = JsonResponse(json)
